refactor: config pkg and diff
