+++ packages/config/package.json
@@ -35,6 +35,15 @@
         "default": "./dist/merge-config.mjs"
       }
     },
+    "./diff": {
+      "source": "./src/diff-config.ts",
+      "types": "./dist/diff-config.d.ts",
+      "require": "./dist/diff-config.js",
+      "import": {
+        "types": "./dist/diff-config.d.mts",
+        "default": "./dist/diff-config.mjs"
+      }
+    },
     "./ts-path": {
       "source": "./src/resolve-ts-path-pattern.ts",
       "types": "./dist/resolve-ts-path-pattern.d.ts",
@@ -65,6 +74,7 @@
     "escalade": "3.1.1",
     "jiti": "^1.19.1",
     "merge-anything": "^5.1.7",
+    "microdiff": "^1.3.2",
     "typescript": "^5.3.3"
   },
   "devDependencies": {
