+++ packages/config/src/index.ts
@@ -1,6 +1,8 @@
-export { findConfigFile } from './find-config'
+export { bundleConfig, type BundleConfigResult } from './bundle-config'
+export { diffConfigs, type DiffConfigResult } from './diff-config'
+export { findConfig } from './find-config'
 export { getConfigDependencies, type GetDepsOptions } from './get-mod-deps'
 export { getResolvedConfig } from './get-resolved-config'
-export { bundleConfigFile, loadConfigFile, resolveConfigFile } from './load-config'
+export { loadConfig } from './load-config'
 export { mergeConfigs } from './merge-config'
 export { convertTsPathsToRegexes } from './ts-config-paths'
