+++ packages/config/src/types.ts
@@ -1,0 +1,4 @@
+export interface ConfigFileOptions {
+  cwd: string
+  file?: string
+}
