+++ packages/node/package.json
@@ -58,7 +58,6 @@
     "lil-fp": "1.4.5",
     "lodash.merge": "4.6.2",
     "look-it-up": "2.1.0",
-    "microdiff": "^1.3.2",
     "outdent": " ^0.8.0",
     "pathe": "^1.1.1",
     "pkg-types": "1.0.3",
