+++ packages/node/src/index.ts
@@ -1,10 +1,10 @@
 export { analyzeTokens, writeAnalyzeJSON } from './analyze-tokens'
 export { Builder } from './builder'
-export { findConfig, loadConfigAndCreateContext } from './config'
+export { codegen } from './codegen'
+export { loadConfigAndCreateContext } from './config'
 export { PandaContext } from './create-context'
 export { cssgen, type CssGenOptions } from './cssgen'
 export { debugFiles } from './debug-files'
-export { codegen } from './codegen'
 export { execCommand } from './exec-command'
 export { extractFile, type CssArtifactType } from './extract'
 export { generate } from './generate'
